% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idar_summaries.R
\name{coef.ida}
\alias{coef.ida}
\title{Get coefficients of SCM estimates}
\usage{
\method{coef}{ida}(object, ...)
}
\arguments{
\item{object}{output of the function \code{estimate}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Get coefficients of SCM estimates
}
