% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacheck_addin.R
\name{datacheck_addin}
\alias{datacheck_addin}
\title{Data checking: check a dataset based on pre-defined information}
\usage{
datacheck_addin()
}
\description{
Data checking: check a dataset based on pre-defined information
}

