% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacheck_createinfo_addin.R
\name{datacheck_createinfo_addin}
\alias{datacheck_createinfo_addin}
\title{Data checking: create a info file for data checking.}
\usage{
datacheck_createinfo_addin()
}
\description{
Data checking: create a info file for data checking.
}

