\name{dummy}
\alias{dummy}
\docType{data}
\title{Dummy Functional Trait Dataset}
\description{
A dummy dataset containing 8 species and 8 functional traits (2 continuous, 2 nominal, 2 ordinal, and 2 binary), with some missing values. Also includes a matrix of species abundances from 10 communities.
}
\usage{
 data(dummy)
 }
\format{
\describe{
\item{trait}{data frame of 8 functional traits on 8 species}
\item{abun}{matrix of abundances of the 8 species from 10 communities}
}
}
\source{
Etienne \enc{Laliberté}{Laliberte} \email{etiennelaliberte@gmail.com}
}
\note{
This dataset has been copied from former package \code{FD} as it was archived on April 15,2026.
}
\keyword{datasets}
