% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idar_estimate.R
\name{estimate}
\alias{estimate}
\title{Estimate DAG model}
\usage{
estimate(
  cause,
  effect,
  dag,
  data,
  categorical = NULL,
  ordinal = NULL,
  weights = NULL
)
}
\arguments{
\item{cause}{a string with the name of the node
to check if its causal effect on 'effect'
is identifiable}

\item{effect}{a string with the name of the node
to check if the causal effect of 'cause'
on this node is identifiable}

\item{dag}{a string with the DAG}

\item{data}{a data.frame with the data}

\item{categorical}{a string vector with the name of the
categorical endogenous variables
(see lavaan documentation for 'ordered'
variables)}

\item{ordinal}{a string vector with the name of the
ordinal discrete endogenous variables
(see lavaan documentation for 'ordered'
variables)}

\item{weights}{string with the name of the variable in the
data frame with the sample weights. Optional.}
}
\description{
Estimate the causal effect based on the DAG model
}
