% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idar_summaries.R
\name{globalfit}
\alias{globalfit}
\title{Global fit measures}
\usage{
globalfit(object, ...)
}
\arguments{
\item{object}{output of the function \code{estimate}}

\item{...}{further arguments passed to or from other methods}
}
\value{
Print the global fit measures for the SCM
estimation. See \code{lavaan::fitmeasures}
for details
}
\description{
Global fit measures
}
