% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idar_identification.R
\name{identification}
\alias{identification}
\title{Check if causal effect if identifiable}
\usage{
identification(cause, effect, dag, plot = FALSE)
}
\arguments{
\item{cause}{a string with the name of the node
to check if its causal effect on 'effect'
is identifiable}

\item{effect}{a string with the name of the node
to check if the causal effect of 'cause'
on this node is identifiable}

\item{dag}{a string with the DAG}

\item{plot}{boolean. If \code{TRUE} plot the DAG and the
identification result}
}
\value{
It returns a \code{dosearch} object from that package
}
\description{
Check if the causal effect is identifiable given the
cause, effect, and the DAG
}
