% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.data.R
\name{inspect.data}
\alias{inspect.data}
\title{To check a dataset}
\usage{
inspect.data(data, info, id, check_missing, plot = FALSE, prefix = "",
  outdir)
}
\arguments{
\item{data}{imported dataset to be checked}

\item{info}{information about variables which will be used to check data}

\item{id}{name of variable that contains subejct identification}

\item{check_missing}{vector of variables to check missingness}

\item{plot}{whether create a plot of distribution of all variables or not}

\item{prefix}{prefix in name of output}

\item{outdir}{directory of output}
}
\value{
.csv file that contains list of potential errors
}
\description{
To check a dataset
}

