% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.each.R
\name{inspect.each}
\alias{inspect.each}
\title{To check a variable}
\usage{
inspect.each(x, varname, type, value = NA, check_missing = TRUE)
}
\arguments{
\item{x}{variable to be checked}

\item{varname}{full name of variable}

\item{type}{type of variable (one of numeric, character, factor, datetime)}

\item{value}{values of factor, range of numeric, format of datetime}

\item{check_missing}{whether check missingness or not}
}
\value{
dataframe contains potential errors
}
\description{
To check a variable
}

