% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idar_summaries.R
\name{localfit}
\alias{localfit}
\title{Local fit measures}
\usage{
localfit(object, ...)
}
\arguments{
\item{object}{output of the function \code{estimate}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Local fit measures
}
\details{
Print the local fit measures for the DAG
and the data set provided. These measures
the correlation between variables (nodes)
for each conditional independence implied
by the DAG.

A good fit means that we do not reject the
hypotheses that the variables are conditionally
independent.
}
