% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idar_plots.R
\name{plot.ida}
\alias{plot.ida}
\title{Plot results of the SCM estimation}
\usage{
\method{plot}{ida}(x, aes_fit = list(), ...)
}
\arguments{
\item{x}{output of the function \code{estimate}}

\item{aes_fit}{a named list with the aesthetic parameters for the
identification expression. See \code{Details}.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot results of the SCM estimation
}
