% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idar_plots.R
\name{plot_dag}
\alias{plot_dag}
\title{Plot the DAG}
\usage{
plot_dag(dag, fit = NULL, aes_fit = list(), show_legend = FALSE)
}
\arguments{
\item{dag}{a string with the DAG}

\item{fit}{output of the function \code{estimate}. If provided
it plots the total effect and the coefficiets of the
SCM alongside the edges of the DAG}

\item{aes_fit}{a named list with the aesthetic parameters for the
identification expression. See \code{Details}.}

\item{show_legend}{boolean. If \code{TRUE}, show legend
with meaning of the nodes aesthetics.}
}
\value{
It returns a ggplot object with the DAG
}
\description{
Plot the DAG
}
