% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idar_plots.R
\name{plot_identification}
\alias{plot_identification}
\title{Plot DAG and identification result}
\usage{
plot_identification(
  cause,
  effect,
  dag,
  scm = FALSE,
  scm_ida = FALSE,
  fit = NULL,
  size_equation = 4,
  size_title = 13,
  aes_scm = list(),
  aes_scm_ida = list(),
  aes_idequation = list(),
  aes_fit = list()
)
}
\arguments{
\item{cause}{a string with the name of the node
to check if its causal effect on 'effect'
is identifiable}

\item{effect}{a string with the name of the node
to check if the causal effect of 'cause'
on this node is identifiable}

\item{dag}{a string with the DAG}

\item{scm}{boolean, if \code{TRUE} plot the SCM
using a linear model for the
conditional dependencies}

\item{scm_ida}{boolean, if \code{TRUE} plot the
identification analysis equations
using linear models}

\item{fit}{output of the function \code{estimate}. If provided
it plots the total effect and the coefficiets of the
SCM alongside the edges of the DAG}

\item{size_equation}{numeric, font size of the expression in the
identification expression subplot}

\item{size_title}{numeric, font size of the title of the
identification expression subplot}

\item{aes_scm}{a named list with the aesthetic parameters for the
SCM equations plot. See \code{Details}.}

\item{aes_scm_ida}{a named list with the aesthetic parameters for the
SCM equations derived from the identification analysis
See \code{Details}.}

\item{aes_idequation}{a named list with the aesthetic parameters for the
identification expression. See \code{Details}.}

\item{aes_fit}{a named list with the aesthetic parameters for the
identification expression. See \code{Details}.}
}
\description{
Plot the DAG and identification result computed using
the \code{identification} function
}
\details{
The named list aes_* contains the aesthetic parameters of different subplots.
The names must match the parameter label.
}
