% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idar_summaries.R
\name{predict.ida}
\alias{predict.ida}
\title{Generate fitted values from the fitted model}
\usage{
\method{predict}{ida}(object, type = "yhat", ...)
}
\arguments{
\item{object}{output of the function \code{estimate}}

\item{type}{a string with the type of output.
If 'yhat,' generate fitted values
of observed nodes. If 'lv,' generate
fitted values for the latent variables.
See \code{lavaan::lavPredict} documentation.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Generate fitted values from the fitted model
}
\details{
It generates fitted values using the data used
to fit the model. See \code{lavaan::lavPredict} documentation.
}
