% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idar_summaries.R
\name{print.ida}
\alias{print.ida}
\title{Print output of fitted model}
\usage{
\method{print}{ida}(x, ...)
}
\arguments{
\item{x}{output of the function \code{estimate}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print output of fitted model
}
