% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idar_summaries.R
\name{residuals.ida}
\alias{residuals.ida}
\title{Return SCM residuals}
\usage{
\method{residuals}{ida}(object, ...)
}
\arguments{
\item{object}{output of the function \code{estimate}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Return SCM residuals
}
\details{
See \code{lavaan::lavResiduals} for details.
}
