% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idar_simulate.R
\name{simulateDAGdata}
\alias{simulateDAGdata}
\title{Simulate data}
\usage{
simulateDAGdata(dag, n = 500, seed = NULL)
}
\arguments{
\item{dag}{a string with the DAG}

\item{n}{sample size}

\item{seed}{number to set the seed}
}
\description{
Simulate data from the DAG
}
