% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idar_summaries.R
\name{summary.ida}
\alias{summary.ida}
\title{Summary of SCM estimates}
\usage{
\method{summary}{ida}(object, ...)
}
\arguments{
\item{object}{output of the function \code{estimate}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary of SCM estimates
}
