% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idar_summaries.R
\name{tidysummary}
\alias{tidysummary}
\title{Full summary of SCM estimated coefficients}
\usage{
tidysummary(fit)
}
\arguments{
\item{fit}{an output of the function \code{estimate}}
}
\value{
The function returns a \code{tibble} object with the
summary of the estimation.
}
\description{
Full summary of SCM estimated coefficients
}
